refactor: removes getAPIPathRoot
realized that the getFileFetchPath function works without getAPIPathRoot appended at the front, so all the API paths should work without this function.
