+++ src/hooks/usePermissionsDialog.ts
@@ -1,11 +1,7 @@
 import React from 'react';
 import toast from 'react-hot-toast';
 
-import {
-  getAPIPathRoot,
-  sendFetchRequest,
-  removeLastSegmentFromPath
-} from '../utils';
+import { sendFetchRequest, removeLastSegmentFromPath } from '../utils';
 import { useCookiesContext } from '../contexts/CookiesContext';
 import type { FileOrFolder } from '../shared.types';
 import { useZoneBrowserContext } from '../contexts/ZoneBrowserContext';
@@ -24,7 +20,7 @@ export default function usePermissionsDialog() {
     try {
       console.log('Change permissions for item:', targetItem);
       await sendFetchRequest(
-        `${getAPIPathRoot()}api/fileglancer/files/${currentFileSharePath?.name}?subpath=${targetItem.path}`,
+        `/api/fileglancer/files/${currentFileSharePath?.name}?subpath=${targetItem.path}`,
         'PATCH',
         cookies['_xsrf'],
         {
